re-structure stats folder, add type annotations for methods and third parameter n to linear congruential and mixed congruential, relocate Hull Dobell check and export as arrow function
