+++ src/stats/generators/gm.ts
