+++ src/stats/generators/linearCongruential.ts
@@ -1,4 +1,4 @@
-import { RandomGeneratorFunc } from "../types";
+import { RandomGeneratorFunc } from "../../types";
 
 interface LinearCongruentialParams{
     a: number, // multiplicador
